requirements.txt
torch==2.6.0
transformers==4.48.2
python-dotenv
